+++ src/main/java/org/retriever/server/dailypet/global/config/security/SecurityConfig.java
@@ -1,6 +1,8 @@
 package org.retriever.server.dailypet.global.config.security;
 
 import lombok.RequiredArgsConstructor;
+import org.retriever.server.dailypet.global.config.jwt.JwtAuthenticationFilter;
+import org.retriever.server.dailypet.global.config.jwt.JwtTokenProvider;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
